requirements.txt
biopython>=1.70
pandas>=0.20.1


stem>=1.8.0
requests>=2.27.1
dryscrape>=1.0

wordcloud>=1.8.1
beautifulsoup4>=4.10.0
cxutils>=0.0.5